+++ app/controllers/admin/sessions_controller.rb
@@ -6,11 +6,11 @@ class Admin::SessionsController < ApplicationController
 
   def create
     user = User.find_by(email: params[:email])
-    if user&.authenticate(params[:password])
+    if user&.authenticate(params[:password]) && user.active?
       session[:user_id] = user.id
       redirect_to admin_tenants_path, notice: "Logged in successfully"
     else
-      flash.now[:alert] = "Invalid email or password"
+      flash.now[:alert] = "Invalid email or password - or user not active"
       render :new, status: :unprocessable_entity
     end
   end